% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_censored_likelihood.R
\name{fit.censoredlikelihoodBR}
\alias{fit.censoredlikelihoodBR}
\title{fit the model based on censored likelihood for Brown-Resnick model}
\usage{
fit.censoredlikelihoodBR(
  obs,
  loc,
  init,
  fixed,
  vario,
  u,
  ST = FALSE,
  maxit = 1000,
  nCores = 1
)
}
\arguments{
\item{obs}{observation list with same marginal distribution, standard pareto distribution}

\item{loc}{location matrix, e.g., longitude and latitude after transformation}

\item{vario}{the variogram function with parameter h,par, and an extra parameter t representing the time position}

\item{u}{threshold}

\item{ST}{logical, default to be FALSE, should the varigoram include a temporal trend}

\item{maxit}{maximum iteration for the optimization}

\item{nCores}{number of cores available for parallelization}

\item{par}{parameters for variogram}
}
\description{
fit the model based on censored likelihood for Brown-Resnick model
}
\examples{
vario <- function(h,par=c(0,3),t=1){ ##return the semi-variogram
## reparametrization
alpha = 2/(1+exp(-par[1]));lambda1 = par[2]
lambda <- exp(lambda1)
val=2*(sqrt(sum(h^2))/lambda)^alpha
return(val)
}
#Define locations
loc <- expand.grid(1:10, 1:10)

#Simulate data
obs <- simulPareto(1000, loc, vario)

#Evaluate risk functional
maxima <- sapply(obs, max)
thres <- quantile(maxima, 0.9)
ncores = detectCores() - 2
#Select exceedances
exceedances <- obs[maxima > thres]
result = fit.censoredlikelihoodBR(obs=exceedances,loc=loc,init=c(0,4),fixed = c(F,F),vario = vario,u = thres,ST = TRUE,nCores = ncores)
}
\author{
Peng Zhong
}
