% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gradientScore.R
\name{fit.gradientScoreBR}
\alias{fit.gradientScoreBR}
\title{fit the model based on censored likelihood for Brown-Resnick model with gradient score method}
\usage{
fit.gradientScoreBR(
  obs,
  loc,
  init,
  fixed,
  vario,
  u,
  ST = FALSE,
  maxit = 1000,
  nCores = 1,
  weightFun = NULL,
  dWeightFun = NULL,
  ...
)
}
\arguments{
\item{obs}{observation list with same marginal distribution, standard pareto distribution}

\item{loc}{location matrix, e.g., longitude and latitude after transformation}

\item{vario}{the variogram function with parameter h,par, and an extra parameter t representing the time position}

\item{u}{threshold}

\item{ST}{logical, default to be FALSE, should the varigoram include a temporal trend}

\item{maxit}{maximum iteration for the optimization}

\item{nCores}{number of cores available for parallelization}

\item{weightFun}{Function of weights.}

\item{dWeightFun}{Partial derivative function of \code{weightFun}.}

\item{par}{parameters for variogram}
}
\description{
fit the model based on censored likelihood for Brown-Resnick model with gradient score method
}
\examples{
vario <- function(h,par=c(0,log(3)),t=1){ ##return the semi-variogram
#reparametrization
alpha = 2/(1+exp(-par[1]));lambda1 = par[2]
lambda <- exp(lambda1)
val=2*(sqrt(sum(h^2))/lambda)^alpha
return(val)
}
#Define locations
loc <- expand.grid(1:10, 1:10)

#Simulate data
obs <- simulPareto(1000, loc, vario)

#Evaluate risk functional
maxima <- sapply(obs, max)
thres <- quantile(maxima, 0.9)
ncores = detectCores() - 2
#Select exceedances
exceedances <- obs[maxima > thres]
result = fit.gradientScoreBR(obs=exceedances,loc=loc,init=c(0,log(4)),fixed = c(F,F),vario = vario,u = thres,ST = TRUE,nCores = ncores)
}
\references{
de Fondeville, R. and Davison A. (2018). High-dimensional peaks-over-threshold inference. Biometrika, 105(3), 575-592.

Wadsworth, J. L. and J. A. Tawn (2014). Efficient inference for spatial extreme value
processes associated to log-Gaussian random functions. Biometrika, 101(1), 1-15.
}
\author{
Peng Zhong
}
